Add persona based views
The persona button on the navbar is used to change the current persona. This is stored in the local storage. The default persona is the "business leader" persona. On the navbar, only the pages that include the current persona will appear. The pages may have same title, but they need to have unique paths.

Touches #8

Release note: None