Update Tagset template (now Class)
Uses the new Hierarchy Tree.
Updated to fetch the hierarchy based on version.
TagSets template is now renamed to Class template.
Updated to show badge for type and version.
Updated to show namespace information as well as relationships that classes have.
Provides version details to the OntologySearchInput

Release note: None