Create OntologyPages
These pages will have a different layout that the Webpages.
They will have a layout similar to an Ontology details page.
The metadata is used to set the version for the tree and the ontology search.
The metadata is also used to make breadcrumbs.
The tree will be open till level 1 for all these pages.
They can contain details about the ontology like structure, root classes or even the changelog of a version.

Release note: None